Specify "pg_start_backup" as BACKUP METHOD in backup_label.
do_pg_backup_start() always returns "streamed" as BACKUP METHOD
in the backup label. But an exclusive backup should use "pg_start_backup"
instead as it was in v14 or before. This commit changes pg_start_backup()
so that it modifies the backup label that do_pg_backup_start() returns
and specifies "pg_start_backup" as BACKUP METHOD in it.