Make installation of pg_exclusive_backup drop built-in backup functions.
Previously if pg_start_backup that this extension provides was executed and
then the PostgreSQL built-in function for non-exclusive bacups,
pg_backup_stop, was executed in the same session, which could make that
pg_backup_stop cause a segmentation fault. To avoid a segmentation fault
in that case, this commit changes "CREATE EXTENSION pg_exclusive_backup"
so that it drops pg_backup_start and pg_backup_stop function in the database
where the command is executed on. Those function remain dropped and
unavailable in that database even after pg_exclusive_backup is removed from
there by DROP EXTENSION command. To use them for non-exclusive backups,
you need to connect to other databases where pg_exclusive_backup has never
been installed on yet.

It was impossible to modify pg_exclusive_backup functions, not to cause
such a segmentation fault, instead of dropping those built-in functions.
Because, to fix the issue, some internal variables in PostgreSQL source code
needed to be changed from outside. But those variables have been exposed yet.